fix: cast SuperPoint desc_map to float32 to prevent grid_sample RuntimeError under AMP
Agent-Logs-Url: https://github.com/MalharRane/XFeat-SuperPoint-Hybrid-Model/sessions/aefbc19f-7cb6-4a41-905f-f319598fcc56

+++ models/hybrid_model.py
@@ -268,7 +268,11 @@ class HybridModel(nn.Module):
                 desc_map = self._sp_desc_map
                 if desc_map is None:
                     raise RuntimeError("SuperPoint hook returned None.")
-        return desc_map
+        # Under torch.amp.autocast the SuperPoint encoder runs in float16, but
+        # F.grid_sample (bicubic) requires input and grid to share dtype.  The
+        # grid is always float32 (derived from keypoints cast with .float()), so
+        # we cast the descriptor map to float32 here to prevent a RuntimeError.
+        return desc_map.float()
 
     # ------------------------------------------------------------------
     # Forward
